requirements.txt
pandas 
numpy 
notebook
scikit-learn
matplotlib
python-box == 6.0.2
pyYAML
tqdm
ensure ==1.0.2
joblib
types-pyYAML
# Flask
# Flask-Cors
# -e .